requirements_base.txt
amqp==1.4.8
anyjson==0.3.3
Babel==2.1.1
billiard==3.3.0.22
boto==2.38.0
celery==3.1.19
colorama==0.3.5
contextlib2==0.4.0
coverage==4.0.3
dj-database-url==0.3.0
Django==1.8.8
django-celery==3.1.17
django-cors-headers==1.1.0
django-countries==3.4.1
django-easymoney==0.5
django-extensions==1.6.1
django-floppyforms==1.6.1
django-idmap==0.4.1
django-inspect-model==0.7
django-sslify==0.2.7
django-sslserver==0.18
django-vanilla-views==1.0.4
djangorestframework==3.3.2
funcsigs==0.4
handy==0.5.3
honcho==0.6.6
IPy==0.83
kombu==3.0.31
mock==1.3.0
ordered-set==1.4.0
otree-core==0.4.31
otree-save-the-change==1.1.3
pbr==1.8.1
pytz==2015.7
raven==5.9.2
requests==2.9.1
selenium==2.48.0
six==1.10.0
wheel==0.24.0
whitenoise==2.0.6
xmltodict==0.9.2
